UBJSON helps some, LZMA seals the deal.
The compression is a LOT better than the LZString derived one, and uses almost entirely libGDX built-in code. The UBJson classes in libGDX are currently broken on GWT but fixable in user code currently (and my PR is already in to fix them for an upcoming release). It might also get reasonably good compression just running LZMA compression on JSON files, or even .dat files.