Ma Shan Zheng is huge, like Go Noto.
So, we now judge if the font is going to be extremely large by checking if the .cmap.txt file would be over 30000 bytes; if it is, we use a 4096x4096 texture and don't even bother trying sizes 280 down to 56. This also maxes out the size it will try at 280 for most fonts, since larger sizes don't scale down very well at all for standard fonts. Moon Dance and Pangolin are also new here!